+++ ATK/Core/TypedBaseFilter.h
@@ -32,9 +32,9 @@ namespace ATK
     
   protected:
     /// This implementation retrieves inputs from other filters and converts it accordingly
-    virtual void process_impl(int size);
+    virtual void process_impl(long size);
     /// Prepares the filter by retrieving the inputs arrays
-    void virtual prepare_process(int size);
+    void virtual prepare_process(long size);
     virtual int get_type() const;
     
     /// Used to convert other filter outputs to DataType*
