+++ ATK/Mock/TriangleCheckerFilter.cpp
@@ -62,12 +62,12 @@ namespace ATK
   }
 
   template<class DataType_>
-  void TriangleCheckerFilter<DataType_>::process_impl(int size)
+  void TriangleCheckerFilter<DataType_>::process_impl(long size)
   {
     
     double real_increment = 2. / input_sampling_rate * frequency;
     
-    for(int i = 0; i < size; ++i)
+    for(long i = 0; i < size; ++i)
     {
       state += real_increment * (ascending ? 1 : -1);
       if(state >= 1)
