+++ ATK/Tools/MSFilter.cpp
@@ -22,7 +22,7 @@ namespace ATK
   }
   
   template<typename DataType_>
-  void MiddleSideFilter<DataType_>::process_impl(int size)
+  void MiddleSideFilter<DataType_>::process_impl(long size)
   {
     if(outputs_size[0] < size)
     {
@@ -35,7 +35,7 @@ namespace ATK
       outputs_size[1] = size;
     }
     
-    for(int i = 0; i < size; ++i)
+    for(long i = 0; i < size; ++i)
     {
       outputs[0][i] = static_cast<DataType>((converted_inputs[0][i] + converted_inputs[1][i]) / 2);
       outputs[1][i] = static_cast<DataType>((converted_inputs[0][i] - converted_inputs[1][i]) / 2);
